// src/platform/packages/shared/kbn-object-versioning/lib/types.d.ts
import type { z } from '@kbn/zod';
import type { Type } from '@kbn/config-schema';
export type Version = number;
export type ObjectTransform<I = unknown, O = unknown> = (input: I) => O;
export interface VersionableObject<UpIn = unknown, UpOut = unknown, DownIn = unknown, DownOut = unknown> {
    /**
     * Schema used to validate a versioned object before up/down transforms.
     * Supports `@kbn/config-schema` or `@kbn/zod` schemas.
     */
    schema?: Type<any> | z.ZodTypeAny;
    down?: ObjectTransform<DownIn, DownOut>;
    up?: ObjectTransform<UpIn, UpOut>;
}
export interface ObjectMigrationDefinition {
    [version: Version]: VersionableObject<any, any, any, any>;
}
export type TransformReturn<T = unknown> = {
    value: T;
    error: null;
} | {
    value: null;
    error: Error;
};
export interface ObjectTransforms<UpIn = unknown, UpOut = unknown, DownIn = unknown, DownOut = unknown> {
    up: <I = UpIn, O = UpOut>(obj: I, version?: Version | 'latest', options?: {
        /** Validate the object _before_ up transform */
        validate?: boolean;
    }) => TransformReturn<O>;
    down: <I = DownIn, O = DownOut>(obj: I, version?: Version | 'latest', options?: {
        /** Validate the object _before_ down transform */
        validate?: boolean;
    }) => TransformReturn<O>;
    validate: (obj: any, version?: Version) => Error | null;
}